NODE-550: Reorganize existing GraphQL schema data types
