+++ node/src/main/scala/io/casperlabs/node/api/graphql/schema/blocks/types/package.scala
@@ -1,33 +1,16 @@
-package io.casperlabs.node.api.graphql
+package io.casperlabs.node.api.graphql.schema.blocks
 
-import java.time.{Instant, ZoneId, ZonedDateTime}
-
-import cats.effect.Effect
-import cats.effect.implicits._
-import cats.implicits._
-import io.casperlabs.blockstorage.BlockStore
-import io.casperlabs.casper.MultiParentCasperRef.MultiParentCasperRef
-import io.casperlabs.casper.SafetyOracle
-import io.casperlabs.casper.api.BlockAPI
-import io.casperlabs.casper.consensus.Block.ProcessedDeploy
+import cats.syntax.either._
+import cats.syntax.option._
+import io.casperlabs.casper.consensus.Block._
+import io.casperlabs.casper.consensus._
 import io.casperlabs.casper.consensus.info.DeployInfo.ProcessingResult
-import io.casperlabs.casper.consensus.info.{BlockInfo, DeployInfo}
-import io.casperlabs.casper.consensus.{Approval, Block, Deploy, Signature}
+import io.casperlabs.casper.consensus.info._
 import io.casperlabs.crypto.codec.{Base16, Base64}
-import io.casperlabs.shared.Log
-import sangria.schema.{fields, _}
+import io.casperlabs.node.api.graphql.schema.utils.DateType
+import sangria.schema._
 
-private[graphql] class GraphQLSchemaBuilder[F[_]: Fs2SubscriptionStream: Log: Effect: MultiParentCasperRef: SafetyOracle: BlockStore: FinalizedBlocksStream] {
-
-  // Not defining inputs yet because it's a read-only field
-  val DateType: ScalarType[Long] = ScalarType[Long](
-    "Date",
-    coerceUserInput = _ => ???,
-    coerceOutput =
-      (l, _) => ZonedDateTime.ofInstant(Instant.ofEpochMilli(l), ZoneId.of("Z")).toString,
-    coerceInput = _ => ???
-  )
-
+package object types {
   val SignatureType = ObjectType(
     "Signature",
     "Signature used to sign Block or Deploy",
@@ -284,120 +267,4 @@ private[graphql] class GraphQLSchemaBuilder[F[_]: Fs2SubscriptionStream: Log: Ef
       )
     )
   )
-
-  val BlockHashPrefix =
-    Argument(
-      "blockHashBase16Prefix",
-      StringType,
-      description = "Prefix or full base-16 hash of a block"
-    )
-
-  val Depth =
-    Argument(
-      "depth",
-      IntType,
-      description = "How many of the top ranks of the DAG to show"
-    )
-
-  val MaxRank =
-    Argument(
-      "maxRank",
-      OptionInputType(LongType),
-      "The maximum rank to to go back from, 0 means go from the current tip of the DAG",
-      0L
-    )
-
-  val DeployHash =
-    Argument(
-      "deployHashBase16",
-      StringType,
-      description = "Base-16 hash of a deploy, must be 64 characters long"
-    )
-
-  val requireFullBlockFields = Set("blockSizeBytes", "deployErrorCount", "deploys")
-
-  def hasAtLeastOne(projections: Vector[ProjectedName], fields: Set[String]): Boolean = {
-    def flatToSet(ps: Vector[ProjectedName], acc: Set[String]): Set[String] =
-      if (ps.isEmpty) {
-        acc
-      } else {
-        val h = ps.head
-        flatToSet(ps.tail, acc + h.name) ++ flatToSet(h.children, acc)
-      }
-
-    flatToSet(projections, Set.empty).intersect(fields).nonEmpty
-  }
-
-  def createSchema: Schema[Unit, Unit] =
-    Schema(
-      query = ObjectType(
-        "Query",
-        fields[Unit, Unit](
-          Field(
-            "block",
-            OptionType(BlockType),
-            arguments = BlockHashPrefix :: Nil,
-            resolve = Projector { (context, projections) =>
-              BlockAPI
-                .getBlockInfoOpt[F](
-                  blockHashBase16 = context.arg(BlockHashPrefix),
-                  full = hasAtLeastOne(projections, requireFullBlockFields)
-                )
-                .toIO
-                .unsafeToFuture()
-            }
-          ),
-          Field(
-            "dagSlice",
-            ListType(BlockType),
-            arguments = Depth :: MaxRank :: Nil,
-            resolve = Projector { (context, projections) =>
-              BlockAPI
-                .getBlockInfosMaybeWithBlocks[F](
-                  depth = context.arg(Depth),
-                  maxRank = context.arg(MaxRank),
-                  full = hasAtLeastOne(projections, requireFullBlockFields)
-                )
-                .toIO
-                .unsafeToFuture()
-            }
-          ),
-          Field(
-            "deploy",
-            OptionType(DeployInfoType),
-            arguments = DeployHash :: Nil,
-            resolve = c => BlockAPI.getDeployInfoOpt[F](c.arg(DeployHash)).toIO.unsafeToFuture()
-          )
-        )
-      ),
-      subscription = ObjectType(
-        "Subscription",
-        fields[Unit, Unit](
-          Field.subs(
-            "finalizedBlocks",
-            BlockType,
-            "Subscribes to new finalized blocks".some,
-            resolve = { c =>
-              // Projectors don't work with Subscriptions
-              val requireFullBlock = c.query.renderCompact
-                .split("[^a-zA-Z0-9]")
-                .collect {
-                  case s if s.trim.nonEmpty => s.trim
-                }
-                .toSet
-                .intersect(requireFullBlockFields)
-                .nonEmpty
-              FinalizedBlocksStream[F].subscribe.evalMap { blockHash =>
-                BlockAPI
-                  .getBlockInfoWithBlock[F](
-                    blockHash = blockHash,
-                    full = requireFullBlock
-                  )
-                  .map(Action(_))
-              }
-            }
-          )
-        )
-      ).some
-    )
 }
