+++ node/src/main/scala/io/casperlabs/node/api/graphql/schema/globalstate/types/package.scala
@@ -1,0 +1,3 @@
+package io.casperlabs.node.api.graphql.schema.globalstate
+
+package object types {}
